+++ src/main/java/com/tanabe/rs/translator/NumberToEnglishTranslator.java
@@ -2,13 +2,15 @@ package com.tanabe.rs.translator;
 
 import com.ibm.icu.text.RuleBasedNumberFormat;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Brian on 7/2/2014.
  */
 public class NumberToEnglishTranslator {
 
 
-    public static String translateNumber(double number){
+    public static String translateNumber(BigDecimal number){
         RuleBasedNumberFormat numberFormat = new RuleBasedNumberFormat(RuleBasedNumberFormat.SPELLOUT);
         return numberFormat.format(number);
     }
