+++ src/test/java/com/tanabe/test/NumberTranslatorTest.java
@@ -2,6 +2,8 @@ package com.tanabe.test;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 import static com.tanabe.rs.translator.NumberToEnglishTranslator.translateNumber;
 import static org.junit.Assert.assertEquals;
 
@@ -12,6 +14,6 @@ public class NumberTranslatorTest {
 
     @Test
     public void testTranslateTwoThousandFiveHundredTwentyThreeAndFourDollars(){
-        assertEquals("Two thousand five hundred twenty-three and 04/100 dollars", translateNumber(2523.04));
+        assertEquals("Two thousand five hundred twenty-three and 04/100 dollars", translateNumber(BigDecimal.valueOf(2523.04)));
     }
 }
